Added AI category

+++ src/main/java/com/tracker/expenses/money/dto/responsedto/ExpenseSummaryDTO.java
@@ -3,11 +3,13 @@ package com.tracker.expenses.money.dto.responsedto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ExpenseSummaryDTO {
-    private double totalSpend;
-    private double monthlySpend;
+    private BigDecimal totalSpend;
+    private BigDecimal monthlySpend;
     private long categoryCount;
     private long transactionCount;
 }
